+++ src/template_cpu.h
@@ -9,7 +9,7 @@
 #define TEMPLATE_CPU_H_
 // #define RAND_MAX 10
 
-void randomI(float *data, int size);
+void randomI(double *data, int size);
